+++ tests/test_auth_token.c
@@ -40,11 +40,19 @@ typedef enum {
   CURLINFO_TOTAL_TIME
 } CURLINFO;
 
+struct token_data test_data;
+
 int curl_easy_perform(CURL *curl)
 {
 	UNUSED(curl);
+	char *msg = "response";
+	int rtn;
+
 	function_called();
-	return (int) mock();
+	rtn = (int) mock();
+	if (0 == rtn)
+	  write_callback_fn (msg, 1, strlen(msg), &test_data);
+	return rtn;
 }
 
 int curl_easy_getinfo(CURL *curl, CURLINFO CURLINFO_RESPONSE_CODE, long response_code)
@@ -67,6 +75,7 @@ void getAuthToken_Null()
 	getAuthToken(&cfg);
 	set_parodus_cfg(&cfg);
 	assert( cfg.client_cert_path == NULL);
+	assert_int_equal (0, (int) cfg.webpa_auth_token[0]);
 }
 
 void getAuthToken_MacNull()
@@ -77,8 +86,10 @@ void getAuthToken_MacNull()
     getAuthToken(&cfg);
     set_parodus_cfg(&cfg);
     assert( cfg.client_cert_path == NULL);
+    assert_int_equal (0, (int) cfg.webpa_auth_token[0]);
 }
 
+#if 0
 void test_requestNewAuthToken_init_fail ()
 {
 	char token[32];
@@ -106,7 +117,7 @@ void test_requestNewAuthToken_init_fail ()
 	assert_int_equal (output, -1);
 	free(cfg.token_server_url);
 }
-
+#endif
 
 void test_requestNewAuthToken_failure ()
 {
@@ -121,6 +132,8 @@ void test_requestNewAuthToken_failure ()
 	parStrncpy(cfg.hw_mac , "123567892366", sizeof(cfg.hw_mac));
 	set_parodus_cfg(&cfg);
 
+	test_data.size = 0;
+	test_data.data = token;
 	will_return (curl_easy_perform, -1);
 	expect_function_calls (curl_easy_perform, 1);
 
@@ -132,6 +145,7 @@ void test_requestNewAuthToken_failure ()
 
 	requestNewAuthToken (token, sizeof(token), 2);
 	assert_int_equal (output, -1);
+	assert_int_equal (0, (int) token[0]);
 	free(cfg.token_server_url);
 }
 
@@ -150,6 +164,8 @@ void test_requestNewAuthToken ()
   parStrncpy(cfg.hw_mac , "123567892366", sizeof(cfg.hw_mac));
   set_parodus_cfg(&cfg);
 
+  test_data.size = 0;
+  test_data.data = token;
   will_return (curl_easy_perform, 0);
   expect_function_calls (curl_easy_perform, 1);
   will_return (curl_easy_getinfo, 0);
@@ -160,6 +176,7 @@ void test_requestNewAuthToken ()
 
   output = requestNewAuthToken (token, sizeof(token), 1);
   assert_int_equal (output, 0);
+  assert_string_equal (token, "response");
   free(cfg.token_server_url);
 }
 
@@ -197,6 +214,8 @@ void test_getAuthToken ()
 	expect_function_calls (curl_easy_getinfo, 1);
 
 	/* To test curl success case */
+	test_data.size = 0;
+	test_data.data = cfg.webpa_auth_token;
 	will_return (curl_easy_perform, 0);
 	expect_function_calls (curl_easy_perform, 1);
 
@@ -207,7 +226,7 @@ void test_getAuthToken ()
 	expect_function_calls (curl_easy_getinfo, 1);
 
 	getAuthToken(&cfg);
-
+	assert_string_equal (cfg.webpa_auth_token, "response");
 	free(cfg.client_cert_path);
 	free(cfg.token_server_url);
 }
@@ -224,6 +243,7 @@ void test_getAuthTokenFailure ()
 	parStrncpy(cfg.hw_serial_number, "Fer23u948590", sizeof(cfg.hw_serial_number));
 	parStrncpy(cfg.hw_mac , "123567892366", sizeof(cfg.hw_mac));
 	set_parodus_cfg(&cfg);
+
 	will_return (curl_easy_perform, -1);
 	expect_function_calls (curl_easy_perform, 1);
 
@@ -261,20 +281,35 @@ void test_getAuthTokenFailure ()
 
 void test_write_callback_fn ()
 {
-	void *buffer;
-	size_t size = 1;
-	size_t nmemb =8;
+	ParodusCfg *cfg;
+	size_t max_data_size = sizeof (cfg->webpa_auth_token);
+	char *buffer1 = "response1";
+	size_t buf1len = strlen(buffer1);
+        char *buffer2 = "R2";
+	size_t buf2len = strlen(buffer2);
+	char buffer3[max_data_size];
 	int out_len=0;
 	struct token_data data;
 	data.size = 0;
-	buffer = strdup("response");
 
-	data.data = (char *) malloc(sizeof(char) * 5);
+	data.data = (char *) malloc(max_data_size);
 	data.data[0] = '\0';
 
-	out_len = write_callback_fn(buffer, size, nmemb, &data);
-	assert_string_equal(data.data, buffer);
-	assert_int_equal( out_len, strlen(buffer));
+	out_len = write_callback_fn(buffer1, 1, buf1len, &data);
+	assert_string_equal(data.data, buffer1);
+	assert_int_equal (out_len, buf1len);
+	assert_int_equal (data.size, buf1len);
+
+	out_len = write_callback_fn(buffer2, 1, buf2len, &data);
+	assert_string_equal(data.data, "response1R2");
+	assert_int_equal (out_len, buf2len);
+	assert_int_equal (data.size, buf1len+buf2len);
+
+	memset (buffer3, 'x', max_data_size);
+	out_len = write_callback_fn(buffer3, 1, max_data_size, &data);
+	assert_int_equal (out_len, 0);
+	assert_int_equal (data.size, 0);
+
 	free(data.data);
 }
 
@@ -285,14 +320,14 @@ void test_write_callback_fn ()
 int main(void)
 {
     const struct CMUnitTest tests[] = {
+	cmocka_unit_test(test_write_callback_fn),
         cmocka_unit_test(test_requestNewAuthToken),
-	cmocka_unit_test(test_requestNewAuthToken_init_fail),
+	// cmocka_unit_test(test_requestNewAuthToken_init_fail),
 	cmocka_unit_test(test_requestNewAuthToken_failure),
 	cmocka_unit_test(getAuthToken_Null),
 	cmocka_unit_test(getAuthToken_MacNull),
 	cmocka_unit_test(test_getAuthToken),
 	cmocka_unit_test(test_getAuthTokenFailure),
-	cmocka_unit_test(test_write_callback_fn),
     };
 
     return cmocka_run_group_tests(tests, NULL, NULL);